Move unit tests from interation_tests direcory
